requirements.txt
pymongo
psycopg2
pika
paramiko
requests
pytest-runner
robotframework
robotframework-sshLibrary
robotframework-seleniumlibrary
robotframework-httplibrary
robotframework-appiumlibrary
robotframework-jsonlibrary
robotframework-mongodblibrary